refactor(servlet): 디렉터리 명 /web/servlet -> /exception 경로로 리팩터링
